+++ grbl/settings.h
@@ -138,7 +138,7 @@ uint32_t get_step_pin_mask(uint8_t i);
 uint32_t get_direction_pin_mask(uint8_t i);
 
 // Returns the limit pin mask according to Grbl's internal axis numbering
-uint8_t get_limit_pin_mask(uint8_t i);
+uint32_t get_limit_pin_mask(uint8_t i);
 
 
 #endif
